+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 
 - Utility test class `AbstractActionTest` and `AbstractEventProviderTest`that eases the definition of action and event provider test classes, respectively.
 - Class `EmptyContentRestHandler` which can be used to define REST endpoints that do not accept any request content (this is for example the case for most of the *GET* requests). 
+- Support for pre/post processors wrapping intent recognition (see [#252](https://github.com/xatkit-bot-platform/xatkit-runtime/issues/252)). This allows to define specific functions to adapt the textual input (e.g. to improve the intent recognition quality), as well as the computed intent (e.g. to normalize values in extracted contexts). **This change breaks the public API**: the `IntentRecognitionProvider` interface is now an abstract class, existing classes implementing the interface should now extend it.
 
 ## Changed
 
@@ -29,6 +30,7 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 - The `XatkitServer` now correctly returns a `404` error if there is no `RestHandler` associated to a requested URI.
 - [#267](https://github.com/xatkit-bot-platform/xatkit-runtime/issues/267): *Change log level of XatkitServer logs to debug*
 - [#269](https://github.com/xatkit-bot-platform/xatkit-runtime/issues/269): *Reduce default logging of DialogFlow API*
+- [#252](https://github.com/xatkit-bot-platform/xatkit-runtime/issues/252): *Pre/Post processing of user inputs*
 
 ## [4.0.0] - 2019-12-01